Integrate two file
